Unit test file indexing methods.
