+++ src/tiledb/cloud/files/__init__.py
@@ -1,8 +1,10 @@
+from . import indexing
 from . import ingestion
 from . import udfs
 from . import utils
 
 __all__ = (
+    "indexing",
     "ingestion",
     "udfs",
     "utils",
