--- src/tiledb/cloud/vector_search/__init__.py
@@ -1,5 +1,0 @@
-from .file_ingestion import ingest
-
-__all__ = [
-    "ingest",
-]
